==> run_scenarios/Results_analysis/3_Validation_field_data_with_sens.Rmd <==
---
title: "Validation_field_data"
author: 'Noémie Huvelle'
date: "2025-12-03"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Prepare R envronment 
```{r load package}
rm(list = ls())
library(ggplot2)
library(stringr)
library(dplyr)
library(tidyverse)
library(patchwork)
library(here)
```
```{r set directories}
graph_file <- here("Graphs/Sens_analysis")
```

```{r load simulation data}
mean_sim_output <- read.csv("Mean_scenarios/mean_ouputs_baseline.csv")
mean_sim_output$DateREP <- str_replace(mean_sim_output$DateREP, " 1", " 2015")
mean_sim_output$DateREP <- as.Date(mean_sim_output$DateREP, "%d %B %Y")
head(mean_sim_output)
```

```{r load field data -- Nest occupied}
#Nesting females  -- Nest occupied
NO_field <- read.csv("Validation_data/NestOccup.csv")
NO_field$date <- as.Date(NO_field$date)
head(NO_field)

field_NO_summarize<-NO_field %>%
  group_by(date, treatment) %>% 
  summarise(female.sum = sum(female))
field_NO_summarize
```
```{r load field data -- Brood cells}
# Cell production
cell_prod_field <- read.csv("Validation_data/CellProd.csv")
cell_prod_field $date <- as.Date(cell_prod_field $date)
head(cell_prod_field)

field_cell_prod_summarize <- cell_prod_field  %>% 
  group_by(date, treatment) %>% 
  summarise(brood.sum = sum(sum.cells))

#field_cell_prod_summarize$brood.sum.cum <- NA

for (i in 1:(nrow(field_cell_prod_summarize))){
  if (i == 1){
    field_cell_prod_summarize$brood.sum.cum[i] <- field_cell_prod_summarize$brood.sum[i]
  } else {
  field_cell_prod_summarize$brood.sum.cum[i] <- field_cell_prod_summarize$brood.sum.cum[i-1] + field_cell_prod_summarize$brood.sum[i]
  }
}


head(field_cell_prod_summarize)
```

```{r load field data -- Offsprinf emergence}
# Offspring emergence
offspring_field <- read.csv("Validation_data/OffspringOverview.csv")
offspring_field$female.mean.weight.g <- NULL
offspring_field$male.mean.weight.g <- NULL
head(offspring_field )

field_sex_ratio <- offspring_field$female.num / ( offspring_field$female.num + offspring_field$male.num)
```

# Plots 
```{r plot brood cell -- sim vs obs}


sim_cell_prod <- mean_sim_output %>%
  filter(variable == "sum.cells")  %>%
  filter(between(DateREP, as.Date("2015-04-24"), as.Date("2015-05-27")))

sim_cell_prod 

p_sum_cell <- ggplot(sim_cell_prod,
       aes(x = DateREP, y = mean)) +
  geom_ribbon(aes(ymin = mean - 2*se, ymax = mean + 2*se),
              alpha = 0.2) +
  geom_line() +
  geom_point(data = field_cell_prod_summarize,
            aes(x = date, y = brood.sum.cum)) +
  geom_line()+
  labs(x = "",
       y = "sum.cells",
       title = "") +
  theme_bw()
p_sum_cell
```


```{r plot nesting females -- sim vs obs}
sim_nest_f <- mean_sim_output %>%
  filter(variable == "bees.nesting.today")  %>%
  filter(between(DateREP, as.Date("2015-04-24"), as.Date("2015-05-27")))
    
sim_nest_f
field_NO_summarize


p_nest_f <- ggplot(sim_nest_f,
       aes(x = DateREP, y = mean)) +
  geom_ribbon(aes(ymin = mean - 2*se, ymax = mean + 2*se),
              alpha = 0.2) +
  geom_line() +
  geom_point(data = field_NO_summarize,
            aes(x = date, y = female.sum)) +
  labs(x = "Date",
       y = "nest.f",
       title = "") +
  theme_bw()

p_nest_f 

```

```{r see male/female ratio -- sim vs obs}

field_sex_ratio <- offspring_field$female.num / ( offspring_field$female.num + offspring_field$male.num)


sim_offspring_male <- mean_sim_output %>%
  filter(variable == "m.emerged.yr")  %>%
  filter(between(DateREP, as.Date("2015-04-24"), as.Date("2015-05-27")))

sim_offspring_female <- mean_sim_output %>%
  filter(variable == "f.emerged.yr")  %>%
  filter(between(DateREP, as.Date("2015-04-24"), as.Date("2015-05-27")))

    
sim_sex_ratio <- sim_offspring_female$mean / (sim_offspring_female$mean + sim_offspring_male$mean)
sim_sex_ratio_df <- data.frame(date = sim_offspring_male$DateREP,
                               sex.ratio = sim_sex_ratio)


p_sex_ratio <- ggplot(sim_sex_ratio_df,
       aes(x = date, y = sex.ratio)) +
  geom_line() +
  labs(x = "",
       y = "Proportion de femelles à l’émergence",
       title = "") +
  geom_point(aes(x=as.Date("2015-05-27") ,y=field_sex_ratio )) +
  theme_bw()

p_sex_ratio 
```

As we can see, the sex ratio is constantly much higher than the observed one (0.25). As the female eggs are produced when there are more foraging ressource, all indicates an overestimation of floral ressources. 

The dip in sex ratio around the 27th of April might be du a lack of ressources around that time (compare with floral quality) OR due to the fact that no new cycle offspring yet ?
```{r combine graphs}
combo_plots <- ( p_sum_cell | p_nest_f | p_sex_ratio) 

combo_plots

ggsave(
  paste(graph_file, "/validation_out_vars.png", sep=""),
  plot   = combo_plots,
  width  = 14,
  height = 4
)
```

# After the sensitivity analysis...
We've seen that the thresholds for precipitations, relative humidity and temperature are the most influent. Let's look a bit at which simulation is the closest to the field data.  The graphs presented before only accounted for the baseline scenario. 

```{r load relevant simulation data}
all_scenarios <- read.csv("mean_outputs_all_scenarios.csv")
str(all_scenarios)
#Create a column which speficies the parameter
all_scenarios <- all_scenarios %>%
  mutate(DateREP =  str_replace(DateREP, " 1", " 2015"),
         DateREP = as.Date(DateREP, format= "%d %B %Y"),
         param_type =
           case_when(
             str_detect(scenario_name, "QN") ~ "lambda.nat",
             str_detect(scenario_name, "T")   ~ "temperature",
             str_detect(scenario_name, "RH")  ~ "humidity",
             TRUE ~ "other"
           ),
         level = case_when(
           str_detect(scenario_name, "baseline") ~ "baseline",
           str_detect(scenario_name, "low") ~ "low",
           str_detect(scenario_name, "high") ~ "high",
           TRUE ~ "other"),
         level = as.factor(level)) %>%
  filter(between(DateREP, as.Date("2015-04-24"), as.Date("2015-05-27")))

#Filter the relevant parameter sand relevant timelines
sim_lambda.nat_sum.cells <- all_scenarios %>%
  filter(param_type == "lambda.nat")  %>%
  filter(variable == "sum.cells")
sim_temp_sum.cells <- all_scenarios %>% filter(param_type == "temperature") %>%
  filter(variable == "sum.cells")
sim_rh_sum.cells <- all_scenarios %>% filter(param_type == "humidity") %>%
  filter(variable == "sum.cells")

base_sim_cell_prod <- sim_cell_prod #Baseline scenario filtered out for relevant dates and relevant variable (sum.cells)
base_sim_cell_prod$level <- as.factor("baseline")
field_cell_prod <- field_cell_prod_summarize %>%
  mutate(sum.cells =  brood.sum.cum)#field data
```

```{r plot legend }
cols_levels <- c(
  "baseline" = "black",
  "low"      = "chartreuse4",
  "high"     = "firebrick"
)

labs_levels <- c(
  baseline = "Baseline",
  low      = "Paramètre -10%",
  high     = "Paramètre +10%"
)

```


```{r plot RH -- sum.cells}
p_RH <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_cell_prod,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_rh_sum.cells,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_cell_prod,
    aes(x = date, y = sum.cells),
    size = 2.5,
    colour = "black"
  ) +
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  theme_bw(base_size = 14) +
  scale_x_date(date_labels = "%d-%m")+
  labs(
    title = "RH.max",
    x = "",
    y = "sum.cells"
  )

p_RH
```

```{r plot T -- sum.cells (plot 2)}
p_T <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_cell_prod,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_temp_sum.cells,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_cell_prod,
    aes(x = date, y = sum.cells),
    size = 2.5,
    colour = "black"
  ) +
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  theme_bw(base_size = 14) +
  scale_x_date(date_labels = "%d-%m")+
  labs(
    title = "T.min",
    x = "Date",
    y = ""
  )

p_T
```

As we can see, if the temperature threshold is set to high (18 + 1.8) then we really get close to the sum.cells observed <3 

```{r  plot alpha -- sum.cells (plot3)}
p_lambda.nat <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_cell_prod,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_lambda.nat_sum.cells,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_cell_prod,
    aes(x = date, y = sum.cells),
    size = 2.5,
    colour = "black"
  ) +
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  scale_x_date(date_labels = "%d-%m")+
  theme_bw(base_size = 14) +
  labs(
    title = "Lambda.nat",
    x = "",
    y = ""
  )

p_lambda.nat
```




```{r plot 3 params}
combo_plots <- ( p_RH | p_T | p_lambda.nat) +
  plot_layout(guides = "collect") &   # légende commune
  theme(
    legend.position  = "bottom",
    legend.direction = "horizontal"
  )

combo_plots

ggsave(
  paste(graph_file, "/sum.cells_3params.png", sep=""),
  plot   = combo_plots,
  width  = 14,
  height = 4
)
```


```{r Calculate RMSE of these 3 parameters}
all_scenarios_sum_cells_params <- all_scenarios %>%
  filter(variable == "sum.cells") %>%
  filter(param_type %in% c("lambda.nat", "temperature", "humidity")) %>%
  select(scenario_name, param_type, level, DateREP, variable, mean)

field_cell_prod_summarize$DateREP <- field_cell_prod_summarize$date

rmse_sens_df <- all_scenarios_sum_cells_params %>%
  inner_join(field_cell_prod_summarize, by = "DateREP") %>%
  group_by(param_type, level, scenario_name) %>%
  summarise(
    n = n(),
    RMSE = sqrt(mean((mean - brood.sum.cum)^2, na.rm = TRUE)),
    .groups = "drop"
  ) %>%
  arrange(param_type, RMSE)

rsme_base_df <- base_sim_cell_prod %>%
  inner_join(field_cell_prod_summarize, by = "DateREP") %>%
  summarise(
    RMSE = sqrt(mean((mean - brood.sum.cum)^2, na.rm = TRUE)),
    n = n()
  )

rmse_sens_df
rsme_base_df 
```

```{r plot RH -- nesting females}
#Filter the relevant parametersand relevant timelines
sim_lambda.nat_nf <- all_scenarios %>%
  filter(param_type == "lambda.nat")  %>%
  filter(variable == "bees.nesting.today")
sim_temp_nf <- all_scenarios %>% filter(param_type == "temperature") %>%
  filter(variable == "bees.nesting.today")
sim_rh_nf <- all_scenarios %>% filter(param_type == "humidity") %>%
  filter(variable == "bees.nesting.today")

base_sim_nf <- sim_nest_f
base_sim_nf$level <- as.factor("baseline")


plot_rh_nf <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_nf,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_rh_nf,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_NO_summarize,
    aes(x = date, y = female.sum),
    size = 2.5,
    colour = "black"
  ) +
  scale_x_date(date_labels = "%d-%m")+
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  theme_bw(base_size = 14) +
  theme(
    legend.position = "bottom"
  ) +
  labs(
    title= "RH.max",
    x = "Date",
    y = "bees.nesting.today"
  )

plot_rh_nf

#ggsave("Images/bees.nesting.today_rh.png", plot_nf,
#     width = 5, height = 3)
```

```{r plot T -- nesting females}
plot_T_nf <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_nf,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_temp_nf,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_NO_summarize,
    aes(x = date, y = female.sum),
    size = 2.5,
    colour = "black"
  ) +
  scale_x_date(date_labels = "%d-%m")+
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  theme_bw(base_size = 14) +
  theme(
    legend.position = "bottom"
  ) +
  labs(
    title= "RH.max",
    x = "Date",
    y = "bees.nesting.today"
  )

plot_T_nf
```

```{r plot lambda.nat -- nesting females}
plot_lambda.nat_nf <- ggplot() +
  geom_line( # baseline simulation
    data = base_sim_nf,
    aes(x = DateREP, y = mean, colour = level),
    alpha = 0.15
  ) +
  geom_line( # sensitivity simulation to a parameter
    data = sim_lambda.nat_nf,
    aes(x = DateREP, y = mean, colour = level),
    linewidth = 1.1
  ) +
  geom_point( # Field observation
    data = field_NO_summarize,
    aes(x = date, y = female.sum),
    size = 2.5,
    colour = "black"
  ) +
  scale_x_date(date_labels = "%d-%m")+
  scale_colour_manual(
    values = cols_levels,
    labels = labs_levels,
    name   = "Scénario"
  ) +
  theme_bw(base_size = 14) +
  theme(
    legend.position = "bottom"
  ) +
  labs(
    title= "RH.max",
    x = "Date",
    y = "bees.nesting.today"
  )

plot_lambda.nat_nf
```
